pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atomix"
version = "0.1.0"
description = "Atomistic Modeling Interface for eXploration - Natural language driven toolkit for ab initio / DFT / atomistic modeling workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "atomix developers"}
]
keywords = [
    "computational chemistry",
    "DFT",
    "VASP",
    "materials science",
    "atomistic simulation",
    "machine learning potentials",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "ase>=3.22.0",
    "numpy>=1.21.0",
    "pymatgen>=2023.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
mlip = [
    "mace-torch>=0.3.0",
    "torch>=2.0.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "atomix[mlip,llm,dev]",
]

[project.scripts]
atomix = "atomix.cli.main:cli"

[project.urls]
Homepage = "https://github.com/atomix-project/atomix"
Documentation = "https://github.com/atomix-project/atomix#readme"
Repository = "https://github.com/atomix-project/atomix"

[tool.setuptools.packages.find]
where = ["."]
include = ["atomix*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
